Added new logic to Super Admin and location

+++ service-api/src/main/java/greencity/dto/GetTariffServiceDto.java
@@ -28,5 +28,5 @@ public class GetTariffServiceDto {
     String createdBy;
     LocalDate editedAt;
     String editedBy;
-    Integer locationId;
+    Long locationId;
 }
